feat: validation middleware

+++ src/controllers/UserController.ts
@@ -1,20 +1,37 @@
 import { Request, Response } from 'express'
 import UserService from '@services/UserService'
+import { validationResult } from 'express-validator'
 
 class UserController {
-	async create(req: Request, res: Response): Promise<void> {
+	async create(req: Request, res: Response) {
 		try {
+			const errors = validationResult(req)
+
+			if (!errors.isEmpty()) {
+				return res.status(400).json({
+					errors: errors.array().map((error) => {
+						return { message: error.msg }
+					}),
+				})
+			}
+
 			await UserService.create(req.body)
 
 			res.status(201).json({ message: 'sucessfully created new user' })
 		} catch (e) {
 			console.log(e)
-			res.status(400).json({ message: 'could not create user' })
+			res.status(400).json({errors: [{ message: e.message}] })
 		}
 	}
 
 	async login(req: Request, res: Response) {
 		try {
+			const errors = validationResult(req)
+
+			if (!errors.isEmpty()) {
+				return res.status(400).json({ errors: errors.array() })
+			}
+
 			const { accessToken, refreshToken } = await UserService.login(req.body)
 
 			res.status(200).json({
@@ -30,6 +47,12 @@ class UserController {
 
 	async refreshToken(req: Request, res: Response) {
 		try {
+			const errors = validationResult(req)
+
+			if (!errors.isEmpty()) {
+				return res.status(400).json({ errors: errors.array() })
+			}
+
 			const { token } = req.body
 
 			const { accessToken, refreshToken } =
@@ -48,6 +71,12 @@ class UserController {
 
 	async logout(req: Request, res: Response) {
 		try {
+			const errors = validationResult(req)
+
+			if (!errors.isEmpty()) {
+				return res.status(400).json({ errors: errors.array() })
+			}
+
 			const { refreshToken } = req.body
 			await UserService.logout(refreshToken)
 			res.status(200).json({ message: 'succesfully logged out' })
@@ -59,6 +88,12 @@ class UserController {
 
 	async recoverPassword(req: Request, res: Response) {
 		try {
+			const errors = validationResult(req)
+
+			if (!errors.isEmpty()) {
+				return res.status(400).json({ errors: errors.array() })
+			}
+
 			const { email } = req.body
 
 			if (!email) throw new Error('email is missing')
@@ -73,25 +108,37 @@ class UserController {
 
 	async changePassword(req: Request, res: Response) {
 		try {
-			const {user, oldPassword, newPassword } = req.body
+			const errors = validationResult(req)
+
+			if (!errors.isEmpty()) {
+				return res.status(400).json({ errors: errors.array() })
+			}
+
+			const { user, oldPassword, newPassword } = req.body
 
 			await UserService.changePassword(user.id, oldPassword, newPassword)
 
-			res.status(200).json({message: 'password changed'})
+			res.status(200).json({ message: 'password changed' })
 		} catch (e) {
 			console.log(e)
-			res.status(400).json({message: e.message})
+			res.status(400).json({ message: e.message })
 		}
 	}
 
-	async delete(req: Request, res: Response){
+	async delete(req: Request, res: Response) {
 		try {
-			const {id} = req.body
+			const errors = validationResult(req)
+
+			if (!errors.isEmpty()) {
+				return res.status(400).json({ errors: errors.array() })
+			}
+
+			const { id } = req.body
 
 			await UserService.delete(id)
 		} catch (e) {
 			console.log(e)
-			res.status(400).json({message: e.message})
+			res.status(400).json({ message: e.message })
 		}
 		res.sendStatus(200)
 	}
